Move old project files into reference directory
* remove unused IDE files
