--- .classpath
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="src" path=""/>
-	<classpathentry kind="lib" path="/Users/taushsampley/Development/java3d-1_5_1-macosx/lib/ext/j3dcore.jar"/>
-	<classpathentry kind="lib" path="/Users/taushsampley/Development/java3d-1_5_1-macosx/lib/ext/j3dutils.jar"/>
-	<classpathentry kind="lib" path="/Users/taushsampley/Development/java3d-1_5_1-macosx/lib/ext/vecmath.jar"/>
-	<classpathentry kind="lib" path="/Users/taushsampley/Development/jogamp-all-platforms/jar/jogl-all.jar"/>
-	<classpathentry kind="output" path=""/>
-</classpath>
