--- Electron.java
@@ -1,6 +1,0 @@
-
-public class Electron {
-	public Electron() {
-
-	}
-}
