+++ old-reference/Electron.java
@@ -1,0 +1,6 @@
+
+public class Electron {
+	public Electron() {
+
+	}
+}
